spam_no_spam_classification/requirements.txt
streamlit
pandas
numpy
plotly.express
scikit-learn
requests